First draft of fight test.
We definitely want to describe the fight in detail.
I'm not sure we want to immediately reconcile the presence of 2 Aliens
in the same city within the same business transactions as the movement.
Can we possibly want the movement to fail because the fight cannot take
place? I guess no, so that could be a warning sign.
You can have a temporary inconsistency moment where you see 2 Aliens in
the same city. I guess it's ok because the fight may take a bit of time
to complete.

There is also some randomness to be sorted out as we don't know who wins
in the fight.