First Event, AliensFightStarted, correctly emitted.
By only substituting the Alien if it's not present in the receiving
city, We can postpone the reconciliation of which Alien will be there
after the fight.
There is an asimmetry: in some cases City::$alien is changed
immediately, in others it is not. Time will tell if this is a wise
choice.