Writing down an AcceptanceTest to automate the flow of Events
The Events from a single transaction must be transformed into Commands
for the next ones until no Commands remain.
The When step is starting to require that we type the Events and we give
some introspection into them instead of using just strings.
The reason is we need to know the current city of an Alien and we can
only project Events for that to work.