Emitting AlienDead
Apparently the trick with __toString() does not work, so I had to update
the assertion for AlienLanded with the Event object. Still, no big deal
as since I'm creating it with scalar data it's easy to create one for
comparison.
Alien classes are a bit of a boilerplate but once I have enough of them
it will be easy to extract a superclass with a dynamic body.